+++ workout-service/src/main/resources/db/migration/V1__init_schema.sql
@@ -20,8 +20,8 @@ CREATE TABLE exercises
 
 CREATE TABLE workout_exercises
 (
-    workout_id  INT REFERENCES workouts (id) ON UPDATE CASCADE ON DELETE CASCADE,
-    exercise_id INT REFERENCES exercises (id) ON UPDATE CASCADE,
+    workout_id  BIGINT REFERENCES workouts (id) ON UPDATE CASCADE ON DELETE CASCADE,
+    exercise_id BIGINT REFERENCES exercises (id) ON UPDATE CASCADE,
     sequence    INT NOT NULL,
     created_at  TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     updated_at  TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
@@ -34,7 +34,7 @@ CREATE INDEX ix_workout_exercises_exercise_id ON workout_exercises (exercise_id)
 CREATE TABLE workout_logs
 (
     id              BIGSERIAL PRIMARY KEY,
-    workout_id      INT REFERENCES workouts (id) ON UPDATE CASCADE ON DELETE CASCADE,
+    workout_id      BIGINT REFERENCES workouts (id) ON UPDATE CASCADE ON DELETE CASCADE,
     start_date_time TIMESTAMP NOT NULL,
     end_date_time   TIMESTAMP NOT NULL,
     created_at      TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
@@ -45,8 +45,8 @@ CREATE INDEX ix_workout_logs_user_id ON workout_logs (workout_id);
 CREATE TABLE workout_log_sets
 (
     id             BIGSERIAL PRIMARY KEY,
-    workout_log_id INT REFERENCES workout_logs (id) ON UPDATE CASCADE ON DELETE CASCADE,
-    exercise_id    INT REFERENCES exercises (id) ON UPDATE CASCADE,
+    workout_log_id BIGINT REFERENCES workout_logs (id) ON UPDATE CASCADE ON DELETE CASCADE,
+    exercise_id    BIGINT REFERENCES exercises (id) ON UPDATE CASCADE,
     set_number     INT NOT NULL,
     repetitions    INT,
     weight         NUMERIC,
